imapclient: add support for X-GM-EXT-1
Add support for X-GM-EXT-1 custom extensions. The extensions are
documented here [0]. They include four primary items:

1. X-GM-RAW
X-GM-RAW passes a quoted string to Gmail's internal search interface.
Modify Search command to accept this parameter.

2. X-GM-MSGID
This exposes the internal Gmail message ID as a uint64 to clients.
Modify Fetch and Search to accept/return this parameter.

3. X-GM-THRID
This exposes a thread ID assigned by Gmail to the message, as a uint64.
Modify Fetch and Search to accept/return this parameter.

4. X-GM-LABELS
This exposes the set of labels applied to a given message. Labels are
equivalent to folders, however it can be advantageous to find all of the
labels/folder a messages belongs to using this extension. Modify Fetch,
Search, and Store commands to work with this extension. Modify Store
signature to accept an interface. This aligns with RFC9051, which does
not define STORE as _only_ accepting a FLAGS data item. FLAGS is the
only defined item in the specification, however it does leave room for
future items to be added.

[0]: https://developers.google.com/gmail/imap/imap-extensions

Signed-off-by: Tim Culverhouse <tim@timculverhouse.com>